move validation to lib file
and update checking to not assume have target_pr (nor indeed a single target variable)
